+++ src/jdk.compiler/share/classes/com/sun/tools/javac/comp/CompileStates.java
@@ -60,8 +60,8 @@ public class CompileStates extends HashMap<Env<AttrContext>, CompileStates.Compi
         FLOW(5),
         TRANSTYPES(6),
         TRANSPATTERNS(7),
-        UNLAMBDA(8),
-        LOWER(9),
+        LOWER(8),
+        UNLAMBDA(9),
         GENERATE(10);
 
         CompileState(int value) {
